Upload file and supporting entities
This commit contains the first throw of the Upload page, where a new
game is added to the library. This page will need refactoring but for
now it is feeding the database with games. Validation is ugly as sin but
is there. The page needs a lot more work but it is functional, thus the
commit.